Refactor wsv2 and async_ipc into providers/persistent
- Update docs for persistent connection provider refactor

Remove sleeps to emulate responses in rapid succession
