+++ docs/conf.py
@@ -88,6 +88,7 @@ exclude_patterns = [
     "web3.providers.rst",
     "web3.providers.rpc.rst",
     "web3.providers.websocket.rst",
+    "web3.providers.persistent.rst",
     "web3.providers.eth_tester.rst",
     "web3.scripts.*",
     "web3.testing.rst",
