+++ web3/__init__.py
@@ -15,8 +15,10 @@ from web3.main import (
     AsyncWeb3,
     Web3,
 )
-from web3.providers.async_ipc import (  # noqa: E402
+from web3.providers.persistent import (  # noqa: E402
     AsyncIPCProvider,
+    PersistentConnectionProvider,
+    WebsocketProviderV2,
 )
 from web3.providers.eth_tester import (  # noqa: E402
     EthereumTesterProvider,
@@ -30,7 +32,6 @@ from web3.providers.rpc import (  # noqa: E402
 )
 from web3.providers.websocket import (  # noqa: E402
     WebsocketProvider,
-    WebsocketProviderV2,
 )
