+++ web3/providers/persistent/__init__.py
@@ -1,0 +1,15 @@
+from .persistent import (
+    PersistentConnectionProvider,
+)
+from .persistent_connection import (
+    PersistentConnection,
+)
+from .request_processor import (
+    RequestProcessor,
+)
+from .async_ipc import (
+    AsyncIPCProvider,
+)
+from .websocket_v2 import (
+    WebsocketProviderV2,
+)
