+++ web3/providers/persistent/websocket_v2.py
