--- web3/providers/websocket/__init__.py
@@ -1,11 +1,0 @@
-from .websocket import (
-    DEFAULT_WEBSOCKET_TIMEOUT,
-    RESTRICTED_WEBSOCKET_KWARGS,
-    WebsocketProvider,
-)
-from .websocket_connection import (
-    PersistentConnection,
-)
-from .websocket_v2 import (
-    WebsocketProviderV2,
-)
